Tile Studio ouput for testing
Changed up some things in Tile Studio's output:

1. Output is now based on original numbering of tiles rather than
stripping out the unused tiles, flattening the layers, and creating a
new renumbered tileset.

2. Outputs number grid for all three layers separately rather than
outputting a single flattened layer.

Collidable map is also still included: 15 is collidable, 0 is
non-collidable.

Graphics for test map are in pictures/graveyard2.0 with proper
numbering. GIVE IT A GO. If it works, I'll go back in and strip out all
the unnecessary information and try to get the formatting to match the
existing .maps perfectly.